PR issues solved
