+++ app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
     <application
-        android:name=".base_app.base_app.BaseApplication"
+        android:name=".base_app.BaseApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -14,7 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/SplashTheme">
         <activity
-            android:name=".base_app.base_app.MainActivity"
+            android:name=".base_app.MainActivity"
             android:exported="true"
             android:screenOrientation="locked">
             <intent-filter>
