+++ app/src/main/java/com/powerhungers/fomezero/base_app/BaseApplication.kt
@@ -1,9 +1,8 @@
-package com.powerhungers.fomezero.base_app.base_app
+package com.powerhungers.fomezero.base_app
 
 import android.app.Application
-import com.powerhungers.fomezero.base_app.feature.login.di.loginModule
+import com.powerhungers.fomezero.feature.login.di.loginModule
 import org.koin.core.context.startKoin
-import org.koin.dsl.module
 
 class BaseApplication : Application() {
 
@@ -11,7 +10,8 @@ class BaseApplication : Application() {
         super.onCreate()
         startKoin {
             modules(
-            loginModule
-        )}
+                loginModule
+            )
+        }
     }
 }
