--- app/src/main/java/com/powerhungers/fomezero/base_app/feature/login/domain/repository/LoginRepository.kt
@@ -1,10 +1,0 @@
-package com.powerhungers.fomezero.base_app.feature.login.domain.repository
-
-import com.powerhungers.fomezero.base_app.feature.login.domain.model.User
-import kotlinx.coroutines.flow.Flow
-
-interface LoginRepository {
-
-    fun login(email: String, password: String): Flow<User>
-
-}
