+++ app/src/main/java/com/powerhungers/fomezero/common/utils/Constants.kt
@@ -1,0 +1,6 @@
+package com.powerhungers.fomezero.common.utils
+
+object Constants {
+    const val EMPTY_EMAIL = "Campo email vazio"
+    const val EMPTY_PASSWORD = "Campo senha vazio"
+}
