+++ app/src/main/java/com/powerhungers/fomezero/feature/login/domain/repository/LoginRepository.kt
@@ -1,0 +1,8 @@
+package com.powerhungers.fomezero.feature.login.domain.repository
+
+import com.powerhungers.fomezero.feature.login.domain.model.User
+import kotlinx.coroutines.flow.Flow
+
+interface LoginRepository {
+    fun login(email: String, password: String): Flow<User>
+}
