+++ app/src/main/java/com/powerhungers/fomezero/feature/login/presentation/LoginFragment.kt
@@ -1,18 +1,17 @@
-package com.powerhungers.fomezero.base_app.feature.login.presentation
+package com.powerhungers.fomezero.feature.login.presentation
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.navigation.fragment.findNavController
 import com.powerhungers.fomezero.databinding.FragmentLoginBinding
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class LoginFragment : Fragment() {
 
     private val binding by lazy { FragmentLoginBinding.inflate(layoutInflater) }
-    private val loginViewModel: LoginViewModel by viewModel()
+    private val viewModel: LoginViewModel by viewModel()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -31,22 +30,20 @@ class LoginFragment : Fragment() {
     private fun handleClickListener() {
         with(binding) {
             btnEnter.setOnClickListener {
-                loginViewModel.login(
+                viewModel.login(
                     editTxtEmail.text.toString(),
-                    editTxtPasswd.text.toString()
+                    editTxtPassword.text.toString()
                 )
             }
         }
     }
 
     private fun addObserver() {
-        loginViewModel.loginLiveData.observe(viewLifecycleOwner) { state ->
+        viewModel.loginLiveData.observe(viewLifecycleOwner) { state ->
             with(binding) {
                 when (state) {
                     is ViewState.Loading -> Unit
-                    is ViewState.Success -> {
-                        findNavController()
-                    }
+                    is ViewState.Success -> Unit
                     is ViewState.Error -> Unit
                     else -> {
                         Unit
