+++ app/src/main/java/com/powerhungers/fomezero/feature/login/presentation/LoginViewModel.kt
@@ -1,18 +1,19 @@
-package com.powerhungers.fomezero.base_app.feature.login.presentation
+package com.powerhungers.fomezero.feature.login.presentation
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.powerhungers.fomezero.base_app.common.extension.postError
-import com.powerhungers.fomezero.base_app.common.extension.postLoading
-import com.powerhungers.fomezero.base_app.common.extension.postNeutral
-import com.powerhungers.fomezero.base_app.common.extension.postSuccess
-import com.powerhungers.fomezero.base_app.feature.login.domain.usecase.LoginUseCase
+import com.powerhungers.fomezero.common.extension.postError
+import com.powerhungers.fomezero.common.extension.postLoading
+import com.powerhungers.fomezero.common.extension.postNeutral
+import com.powerhungers.fomezero.common.extension.postSuccess
+import com.powerhungers.fomezero.feature.login.domain.usecase.LoginUseCase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.launch
 
 class LoginViewModel(private val loginUseCase: LoginUseCase) : ViewModel() {
@@ -26,9 +27,9 @@ class LoginViewModel(private val loginUseCase: LoginUseCase) : ViewModel() {
 
             loginUseCase(email, password)
                 .flowOn(Dispatchers.IO)
-                .catch {
-                    loginViewState.postError(it)
-                }.collect { loginViewState.postSuccess(Unit) }
+                .onStart { loginViewState.postLoading() }
+                .catch { loginViewState.postError(it) }
+                .collect { loginViewState.postSuccess(Unit) }
         }
     }
