+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.powerhungers.fomezero.base_app.feature.login.presentation.LoginFragment"
+        android:name="com.powerhungers.fomezero.feature.login.presentation.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login" />
 </navigation>
